refactor(backends): decompose codex.py into 4 modules
