+++ tests/arch/test_subpackage_isolation.py
@@ -1297,59 +1297,20 @@ _LINE_LIMIT_EXEMPTIONS: dict[str, tuple[int, str]] = {
         "run_skill launch denial paths before command construction (+139 net lines)",
     ),
     "execution/backends/codex.py": (
-        2444,
-        "REQ-CNST-010-E9: Codex backend — skill_sigil capability threading adds multi-line "
-        "keyword args to _ensure_skill_prefix call sites and _has_prefix guard; "
-        "write_guard_tool_names env injection adds 7 lines to _codex_exec_extras; "
-        "session_meta NDJSON support and process_name_aliases add ~8 lines; "
-        "explicit plugin_install_capable + supports_context_window_suffix kwargs for arch guard; "
-        "CodexSessionLocator nominally subclasses SessionLocator Protocol with codex_home "
-        "promoted from locate_session parameter to frozen dataclass field (3 net lines: "
-        "field declaration, blank line between field and method, SessionLocator in import block)"
-        "; project_log_dir method added to CodexSessionLocator (+3 net lines)"
-        "; session_log_path method added to CodexSessionLocator (+5 net lines)"
-        "; process_idle_timeout_ms field wired through build_skill_session_cmd and "
-        "build_food_truck_cmd CmdSpec constructors (+8 net lines)"
-        "; CapabilityNotSupportedError capability-gate in build_inspector_cmd (+1 net line); "
-        "AGENT_BACKEND_ENV_VAR injection in build_interactive_cmd merged_extras (+1 net line) "
-        "and build_resume_cmd _codex_exec_extras call expansion to multi-line (+2 net lines) "
-        "for T5-P4-A3-WP3 guard-hook backend dispatch"
-        "; #4585 excludes reader-only roles from ordinary Codex agent materialization (+1 line)"
-        "; _materialize_profile_skills function (~43 lines) for T5-P4-A4-WP2 profile skill "
-        "materialization into Codex session directories; registered-daemon launch identity "
-        "forwarding adds both required environment keys to all Codex command shapes (+8 net lines)"
-        "; debug-level symlink failure log in _materialize_profile_skills (+5 net lines)"
-        "; env-assembly consolidation via _assemble_shared_env_extras (T5-P4-A1-WP2)"
-        "; fleet inspector child-boundary defaults across Codex command builders (+5 net lines)"
-        "; explicit parameter dispositions for "
-        "plugin_source/output_format/exit_after_stop_delay_ms "
-        "replacing noqa:F841 silent discards (+18 net lines) for T5-P4-A2-WP1"
-        "; github_api_callable field + evidence comment in BackendCapabilities (+2 net lines)"
-        "; output-discipline delivery for fresh interactive sessions and generated agent "
-        "TOMLs (+12 net lines)"
-        "; _register_agent_tomls session-config registration for generated roles "
-        "(+39 net lines)"
-        "; REQ-SEM-ADAPT-001 semantic-plan adaptation remains on the registered Codex "
-        "backend so model and reasoning policy resolution has one authority; "
-        "interactive Codex startup validation, explicit generated-home construction, "
-        "profile probing, and durable cook-storage adapter integration remain co-located "
-        "with the backend whose command grammar they validate; managed native-shell "
-        "decision and lineage-reference injection remain adjacent to the Codex command "
-        "builders that own the protected environment boundary; #4443 adds canonical "
-        "agent-definition projection, parent/child sandbox precedence, and specialized "
-        "Codex explorer role registration and invocation wiring; #4478 review "
-        "remediation: build_skill_session_cmd/build_resume_cmd gain an "
-        "include_scope_discipline parameter and build_interactive_cmd's suffix call is "
-        "widened to codex_discipline_suffix(include_scope=True) so scope-discipline "
-        "delivery scoping stays adjacent to the same command builders that already own "
-        "prompt-injection composition (+14 net lines)"
-        "; #4488/#4489/#4492 explorer surface authority: setup_session_dir gains "
-        "explorer-role TOML exclusion filter for unbound sessions (+7 net lines) and "
-        "explicit session-scoped capability disposition (+1 net line); #4507 adds runtime "
-        "child cardinality rendering and protects the live-web bundled role (+12 net lines)"
-        "; #4566 pins ORCHESTRATOR file auth, durable auth linkage, and role-exact profile "
-        "materialization at the backend-owned generated-home setup boundary (+42 net lines); "
-        "launch/state MCP forwarding defaults remain in the Codex env policy (+2 net lines)",
+        1300,
+        "REQ-CNST-010-E9-narrowed: CodexBackend class alone is 1037 lines (cmd/cmd-spec "
+        "grammar with build_skill_session_cmd/build_food_truck_cmd/build_interactive_cmd/validate_interactive_invocation/setup_session_dir), "
+        "with the four cmd-builder methods (≈465 lines) tightly coupled to CodexBackend "
+        "state. After #4664 decomposition, command vocabulary, env policy, state-readiness "
+        "probe, and session locator moved to _codex_cmd_builders.py; probes to "
+        "_codex_probes.py; explorer projection to _codex_explorer_projection.py. "
+        "CodexBackend retains all five cmd-builder methods because each touches "
+        "instance state (capabilities, env policy, flag vocabulary, session locator) "
+        "and the cmd-spec grammar is the backend's authority boundary — splitting these "
+        "would force a separate mutable state object and break the protocol. The "
+        "remaining slimmed CodexBackend is 1211 lines; cap raised to 1300 to acknowledge "
+        "the architectural seam that the decomposition could not cross without breaking "
+        "the backend dataclass invariant.",
     ),
     "execution/backends/claude.py": (
         1250,
